# PERF-#4560: Implement '.corr()' method using MapReduce pattern