<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/development/contributing.html
if you have questions about contributing.
-->

## What do these changes do?

This PR implements the `.corr()` method using the MapReduce approach, where at the Map phase we compute the required aggregations for the correlation matrix (pairwise multiplication, sum, sum of squares, number of observations) and then at the Reduce phase we build a correlation matrix itself using these values.

The most challenging part of this PR was dealing with NaN values, as in this case we have to exclude the NaN parts of each sum/square of sums/counts for each column individually. This makes the case where there are NaN values in the dataframe the least performant in Modin, and the hardest to understand. However, I tried my bests to put as much in-code comments as possible to explain the procedure that we follow in case of MapReduce implemented correlation that consider NaN values.

Here's the performance of this new implementation measured on a 112 cores Intel(R) Xeon(R) Gold 6238R CPU @ 2.20GHz with different `num_cpus`:

(benchmark mode disables laziness for the operations)
`MODIN_BENCHMARK_MODE=TRUE MODIN_CPUS=112`:

| | pandas | modin on this PR | modin master |
|-|-|-|-|
| [test case from the issue](https://github.com/modin-project/modin/issues/4560#issuecomment-1151498644) (524_288, 100), <br> pure int dataset | 10.55s | 1.75s | 40.21s |
| [test case from the issue](https://github.com/modin-project/modin/issues/4560#issuecomment-1151498644) (524_288, 100), <br> two NaN values in the dataset:  | 10.53s | 2.21s | 40.67s |
| [test case from the issue](https://github.com/modin-project/modin/issues/4560#issuecomment-1151498644) (524_288, 100), <br> 20% of the dataset values are NaNs:  | 10.52s | 3.76s | 44.05s |
| real-world dataset from our internal <br> workload (1_552_210, 214), 77% of the <br> dataset values are NaNs | 67.71s | 35.74s | 75.37s |
| real-world dataset from our internal <br> workload (1_552_210, 214), but also <br> calling `.fillna(method="ffill")` <br> to get rid of NaN values | 140.34s | 10.19s | 200+s |


`MODIN_BENCHMARK_MODE=TRUE MODIN_CPUS=16`:

| | pandas | modin on this PR | modin master |
|-|-|-|-|
| [test case from the issue](https://github.com/modin-project/modin/issues/4560#issuecomment-1151498644) (524_288, 100), <br> pure int dataset | 10.55s | 1.51s | 39.95s |
| [test case from the issue](https://github.com/modin-project/modin/issues/4560#issuecomment-1151498644) (524_288, 100), <br> two NaN values in the dataset:  | 10.53s | 3.41s | 39.01s |
| [test case from the issue](https://github.com/modin-project/modin/issues/4560#issuecomment-1151498644) (524_288, 100), <br> 20% of the dataset values are NaNs:  | 10.52s | 4.07s | 38.06s |
| real-world dataset from our internal <br> workload (1_552_210, 214), 77% of the <br> dataset values are NaNs | 67.71s | 54.46s | 68.39s |
| real-world dataset from our internal <br> workload (1_552_210, 214), but also <br> calling `.fillna(method="ffill")` <br> to get rid of NaN values | 140.34s | 13.44s | 200+s |

<!-- Please give a short brief about these changes. -->

- [x] first commit message and PR title follow format outlined [here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
  > **_NOTE:_**  If you edit the PR title to match this format, you need to add another commit (even if it's empty) or amend your last commit for the CI job that checks the PR title to pick up the new PR title.
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #4560  <!-- issue must be created for each patch -->
- [x] tests added and passing
- [x] module layout described at `docs/development/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->